requirements.txt
fastapi==0.109.1
uvicorn[standard]==0.27.1
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
python-dotenv==1.0.1
pydantic-settings==2.1.0
redis==5.0.1
alembic==1.12.1

# Testing
pytest==7.4.4
httpx==0.26.0
pytest-asyncio==0.23.2
